:rafactor:moved all types to a separate module
