--- src/core/components/index.html
@@ -1,4 +1,0 @@
-<div class="tickets-user__name"></div>
-<div class="tickets-user__date"></div>
-<div class="tickets-user__amount"></div>
-<div class="tickets-user__price"></div>
